2-语料预处理Agent/8-QA算子/data-engine-dev/data-engine-dev/zzj/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
boto3==1.39.10
botocore==1.39.10
certifi==2025.7.14
charset-normalizer==3.4.2
distro==1.9.0
EbookLib==0.19
et_xmlfile==2.0.0
exceptiongroup==1.3.0
h11==0.16.0
html2text==2025.4.15
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
loguru==0.6.0
lxml==6.0.0
mobi==0.3.3
numpy==2.2.6
openai==1.84.0
openpyxl==3.1.5
pandas==2.2.3
pdf2image==1.17.0
Pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
PyMuPDF==1.26.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-pptx==1.0.2
pytz==2025.2
requests==2.32.4
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tqdm==4.67.1
typing_extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
xlrd==2.0.2
xlsxwriter==3.2.5